requirements.txt
anyio==3.5.0
bidict==0.21.4
click==8.0.3
colorama==0.4.4
fastapi==0.73.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SocketIO==5.1.1
gunicorn==20.1.0
idna==3.3
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.22.2
opencv-python==4.5.5.62
Pillow==9.0.1
pydantic==1.9.0
pyshine==0.0.9
python-engineio==4.3.1
python-socketio==5.5.2
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
torch>=1.7.0
torchvision>=0.8.1
Werkzeug==2.0.3
